[refactor] 사용자 xp 추가 기능 구현
