+++ src/main/kotlin/com/meokq/api/challenge/controller/ChallengeController.kt
@@ -11,10 +11,10 @@ import com.meokq.api.core.ResponseEntityCreation
 import com.meokq.api.core.dto.BaseListRespV2
 import com.meokq.api.core.dto.BaseResp
 import com.meokq.api.core.enums.ErrorStatus
-import com.meokq.api.logs.annotations.GrantXp
-import com.meokq.api.logs.annotations.ReturnXp
-import com.meokq.api.logs.processor.impl.ChallengeXpProcessorImpl
-import com.meokq.api.logs.processor.impl.ChallengeXpReturnProcessorImpl
+import com.meokq.api.xp.annotations.GrantXp
+import com.meokq.api.xp.annotations.ReturnXp
+import com.meokq.api.xp.processor.impl.ChallengeXpProcessorImpl
+import com.meokq.api.xp.processor.impl.ChallengeXpReturnProcessorImpl
 import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.validation.Valid
 import org.springframework.data.domain.Page
