+++ src/main/kotlin/com/meokq/api/xp/annotations/ExplainSelectXpList.kt
@@ -1,4 +1,4 @@
-package com.meokq.api.logs.annotations
+package com.meokq.api.xp.annotations
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.media.Content
