+++ src/main/kotlin/com/meokq/api/xp/annotations/ReturnXp.kt
@@ -1,6 +1,6 @@
-package com.meokq.api.logs.annotations
+package com.meokq.api.xp.annotations
 
-import com.meokq.api.logs.processor.XpReturnProcessor
+import com.meokq.api.xp.processor.XpReturnProcessor
 import java.lang.annotation.ElementType
 import java.lang.annotation.Target
 import kotlin.reflect.KClass
