+++ src/main/kotlin/com/meokq/api/xp/processor/GrantXpAspect.kt
@@ -1,9 +1,9 @@
-package com.meokq.api.logs.processor
+package com.meokq.api.xp.processor
 
 import com.meokq.api.core.AuthDataProvider
 import com.meokq.api.user.service.CustomerService
-import com.meokq.api.logs.annotations.GrantXp
-import com.meokq.api.logs.annotations.ReturnXp
+import com.meokq.api.xp.annotations.GrantXp
+import com.meokq.api.xp.annotations.ReturnXp
 import org.aspectj.lang.ProceedingJoinPoint
 import org.aspectj.lang.annotation.Around
 import org.aspectj.lang.annotation.Aspect
