+++ src/main/kotlin/com/meokq/api/xp/processor/XpReturnProcessor.kt
@@ -1,4 +1,4 @@
-package com.meokq.api.logs.processor
+package com.meokq.api.xp.processor
 
 import com.meokq.api.user.request.CustomerXpReq
