+++ src/main/kotlin/com/meokq/api/xp/processor/impl/EmojiXpProcessorImpl.kt
@@ -1,8 +1,8 @@
-package com.meokq.api.logs.processor.impl
+package com.meokq.api.xp.processor.impl
 
 import com.meokq.api.user.request.CustomerXpReq
-import com.meokq.api.logs.processor.UserAction
-import com.meokq.api.logs.processor.XpProcessor
+import com.meokq.api.xp.processor.UserAction
+import com.meokq.api.xp.processor.XpProcessor
 import org.springframework.stereotype.Component
 
 @Component
