+++ src/main/kotlin/com/meokq/api/xp/repository/XpHisRepository.kt
@@ -1,7 +1,7 @@
-package com.meokq.api.logs.repository
+package com.meokq.api.xp.repository
 
 import com.meokq.api.core.repository.BaseRepository
-import com.meokq.api.logs.model.XpHistory
+import com.meokq.api.xp.model.XpHistory
 
 interface XpHisRepository: BaseRepository<XpHistory, String> {
     fun deleteByTitle(title :String)
