+++ src/main/kotlin/com/meokq/api/xp/repository/XpHisSpecification.kt
@@ -1,8 +1,8 @@
-package com.meokq.api.logs.repository
+package com.meokq.api.xp.repository
 
 import com.meokq.api.core.specification.BaseSpecificationV2
 import com.meokq.api.core.specification.SpecificationDto
-import com.meokq.api.logs.model.XpHistory
+import com.meokq.api.xp.model.XpHistory
 
 object XpHisSpecification: BaseSpecificationV2<XpHistory> {
