+++ src/test/kotlin/com/meokq/api/xp/XpHisServiceTest.kt
@@ -1,11 +1,11 @@
-package com.meokq.api.logs
+package com.meokq.api.xp
 
 import com.meokq.api.TestData
 import com.meokq.api.user.model.Customer
 import com.meokq.api.user.service.CustomerService
-import com.meokq.api.logs.dto.XpSearchDto
-import com.meokq.api.logs.model.XpHistory
-import com.meokq.api.logs.service.XpHisService
+import com.meokq.api.xp.dto.XpSearchDto
+import com.meokq.api.xp.model.XpHistory
+import com.meokq.api.xp.service.XpHisService
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
